more support for long vectors
git-svn-id: https://svn.r-project.org/R/trunk@60241 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/main/coerce.c
@@ -2331,7 +2331,7 @@ SEXP attribute_hidden do_docall(SEXP call, SEXP op, SEXP args, SEXP rho)
 	SETCAR(c, mkPROMISE(VECTOR_ELT(args, i), rho));
 	SET_PRVALUE(CAR(c), VECTOR_ELT(args, i)); */
 #endif
-	if (ItemName(names, i) != R_NilValue)
+	    if (ItemName(names, (int)i) != R_NilValue)
 	    SET_TAG(c, install(translateChar(ItemName(names, i))));
 	c = CDR(c);
     }
